Refactor PubStyle tree to remove  use of Router
React Router was being used to render children of PubStyle component
tree, to transfer data between pages, and to mount components.
It was not possible to use URLs to navigate the app, however,
and so the app was refactored so that data is passed via context and
the job of conditionally mounting components is done via component.